delete portfolio
